# Phase_2/specs/003-todo-backend-db/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: Todo Backend API
  version: 1.0.0
  description: API for managing todo tasks with persistent database storage and multi-user isolation.

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

security:
  - bearerAuth: []

paths:
  /todos:
    get:
      summary: List authenticated user's tasks
      responses:
        200:
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        401:
          description: Unauthorized
    post:
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        201:
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        400:
          description: Validation error

  /todos/{id}:
    patch:
      summary: Update or toggle a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        200:
          description: Task updated
        404:
          description: Task not found or not owned by user
    delete:
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        204:
          description: Task deleted
        404:
          description: Task not found or not owned by user

components:
  schemas:
    Task:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        completed: { type: boolean }
        user_id: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    TaskCreate:
      type: object
      required: [title]
      properties:
        title: { type: string }
        description: { type: string }
    TaskUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        completed: { type: boolean }